Example of Level Select
First level select example. Arrows to switch between screens, title of
map. Birds eye view of the whole map not in detail plus a detailed
"photo" of life in the area and hint of boss.  Includes a brief
description of the area plus an objective (gives the player more of a
reason to kill). So I was thinking player kill boss to retrieve
important object, etc.